moved components to their own file
